re-import of pygtk2 as py-gtk2 This was requested by Lubomir Sedlacik <salo@Xtrmntr.org> to match pkgsrc's naming scheme for Python pkgs.
This archive contains modules that allow you to use gtk in Python
programs.  At present, it is a fairly complete set of bindings.
Despite the low version number, this piece of software is quite
useful, and is usable to write moderately complex programs.  (see the
examples directory for some examples of the simpler programs you could
write).