re-import of pyPgSQL as py-PgSQL This was requested by Lubomir Sedlacik <salo@Xtrmntr.org> to match pkgsrc's naming scheme for Python pkgs.
pyPgSQL is a package of two modules that provide a Python DB-API 2.0 compliant
interface to PostgreSQL databases. The first module, libpq, exports the
PostgreSQL C API to Python. This module is written in C and can be compiled
into Python or can be dynamically loaded on demand. The second module, PgSQL,
provides the DB-API 2.0 compliant interface and support for various PostgreSQL
data types, such as INT8, NUMERIC, MONEY, BOOL, ARRAYS, etc. This module is
written in Python.