Initial import of the allegro library, version 4.1.8: Allegro is a cross-platform library intended for use in computer games and other types of multimedia programming.
According to the Oxford Companion to Music, Allegro is the Italian for
"quick, lively, bright". It is also a recursive acronym which stands for
"Allegro Low Level Game Routines".

A wide range of extension packages and add-on modules are also available,
which can be found in the "Library Extensions" section of the Allegro.cc
website, http://www.allegro.cc/.