Initial import of liquidwar, version 5.5.8. Doesn't work, sadly. Help welcome.
Liquid War is a unique multiplayer wargame. Its rules are truly original
and have been invented by Thomas Colcombet. You control an army of liquid
and have to try and eat your opponents. A single player mode is available,
but the game is definitely designed to be multiplayer, and has network
support.