This is Sun's Java[tm] Development Kit, version 1.2.2, made buildable and usable natively on NetBSD/i386 by Greg Lewis and a host of others.
===== The ANNOUNCEMENT from the patchset:

This is a new set of BETA source code patches for the JDK 1.2.2 source.
These patches are unofficial and no abuse should be directed towards the
BSD Java Porting Team for them :).  This new set (patchset 11) can be
gotten from http://www.eyesbeyond.com/freebsddom/java/jdk.html (this
message is included in the downloadable tar ball as the file ANNOUNCEMENT).

Please read the file THANKS for a list of the large number of people who
should be thanked that things have gotten this far!

The main focus of this release is as follows:

1. Support for NetBSD (Scott Bartram)

2. Support for OpenBSD (Matt Behrens, Christian Gruber).

3. JPDA support for 1.2.2.