This is Sun's Java[tm] Development Kit, version 1.3.1, made buildable and usable natively on NetBSD/i386 by Greg Lewis and a host of others.
===== From the patchset README:

This is a set of BETA source code patches for the JDK 1.3.1 SCSL source.
It is sincerely hoped that this patchset is suitable for general use,
although no warranty for such use is expressed or implied.  Use at your
own risk.  The patchset itself is available for download from
http://www.eyesbeyond.com/freebsddom/java/jdk13.html (this message is
included in the downloadable tar ball as the file README).

Please read the file THANKS for attributions.

The JDK built with these patches can only be considered of beta quality
when using the default green threads subsystem with the Classic VM.
These settings are the default.  It is not recommended that this JDK
be placed into production use with other settings (e.g., native threads,
HotSpot).