Initial import of gtranslator, version 0.43.
Gtranslator is a translation-supporting program for the Gnome project.
Simply speaking, it is a helper for developing ".pot" files used with the
GNU Gettext library to provide i18n and l10n features.  It is offering
little functionality, but it will be more useful and complex later on.

Provided in PR pkg/19794 by Juan RP.  The package is *not* reviewed nor
tested yet...