Import of gltron version 0.62.
You steer a futuristic bike, called lightcycle.
Combat takes place in a rectangular arena. Your
bike leaves a trail behind, which is like a wall.
The goal is to force the other players to drive
into a wall. The winner is the last player alive.

(Besides, every geek in this world knows the film.)