Initial import.
qinx is  widget style and window decoration for KDE. It is inspired by
the QNX Photon desktop.

WWW: http://www.usermode.org/code.html

This was supplied in pkg/18055 by Miguel Mendez <flynn@energyhq.homeip.net>.
The PR can be closed as soon as this pkg gets imported to pkgsrc.