Initial import.
VCDGear is a program designed to allow a user to extract MPEG streams from
CD images, convert VCD files to MPEG, correct MPEG errors, and more -- all
in a single step.  Initially developed back in late 1997, the program has
grown to do various extractions, conversions, and corrections on the fly.
Cross-platform support will allow different machines to process and
generate output that is compatible between one another.

People requiring the need to transport their video materials will find
VCDGear as a useful tool to assist them in getting the video to play on
their computer or Video CD player.

This was supplied in pkg/19065 by <collver1@attbi.com>.
The PR can be closed as soon as this gets moved to pkgsrc.