Initial import of ap-bandwidth, version 2.0.3:
mod_bandwidth is a module for the Apache webserver that allows setting of
server-wide or per connection bandwidth limits, based on the directory,
size of files and remote IP/domain.

Provided in PR pkg/18289 by Todd Vierling.  I'm not familiar enough with
apache modules to see if this package is completely correct.