Import of mplayerplug-in 0.71
mplayerplug-in is a browser plugin that uses mplayer to play videos
from websites. A current version of mplayer (0.90rc4) is required. 
This plugin gives Netscape, Mozilla and Konqueror the ability to 
play videos off the net.

This package depends on Mozilla and Mplayer to work.
Also can works with konqueror, but I prefer Mozilla as 
default browser.