Import of streamripper 1.32
Using information which a SHOUTcast server can optionally send,
this program breaks an audio stream into "tracks," stores them in
separate files as they arrive, and names the files by appending
".mp3" to the name of the track.  The tracks can be listened to at
the user's leisure with an MPEG Layer 3 audio player.