Import of directfb 0.9.17
DirectFB is a thin library that provides hardware graphics acceleration, 
input device handling and abstraction, integrated windowing system with 
support for translucent windows and multiple display layers on top of the
Linux Framebuffer Device.
It is a complete hardware abstraction layer with software fallbacks for 
every graphics operation that is not supported by the underlying hardware.
DirectFB adds graphical power to embedded systems and sets a new standard 
for graphics under Linux and BSD ( with SDL ).