Import of lpe 1.2.6.
LPE is a small, efficient programmer's editor for UNIX systems. It has grown
from the result of a single night of hacking into a very capable and very 
versatile editor that defines it's own philosophy opf text editing. That 
philosophy is that awk and sed are the right tools for most advanced editing 
jobs, and a text editor should make simple operations easy and make text look 
good.

http://cdsmith.twu.net/professional/opensource/lpe.html

Closes my own pr pkg/19569.