Update gnomeicu to 0.99.
GnomeICU is a clone of Mirabilis' popular ICQ written with GTK.
The original source was taken from Matt Smith's mICQ. This is ment as
a replacement for the JavaICQ, which is slow and buggy. If you would
like to contribute, please contact Jeremy Wise ( jwise@pathwaynet.com ).
You may also want to visit The ICQ Protocol Site at:
[ http://www.student.nada.kth.se/~d95-mih/icq/ ] for information about
the ICQ protocol and mailing list.

This is the GNOME2 port of gnomeicu. I wasn't sure whether we should replace
the gnomeicu in pkgsrc with this version, or we should keep the GNOME1 version
around for people w/o GNOME2. In the latter case we could rename this pkg
to gnomeicu2. In any case this requires testing first, as I personally only
use it to test the system tray support.