Initial import of Windowlab 1.09
WindowLab is a small and simple window manager of novel design.

It's click-to-focus, shares its window depth policy with the Amiga, 
and has a window resizing/reshaping method similar to that of 8/5 
from Plan 9. WindowLab maintains the illusion of direct manipulation 
by constraining the mouse pointer when appropriate, ie when a window 
cannot be dragged any further in one direction. The pointer is also 
constrained vertically (effectively making the target menu items 
infinitely tall according to Fitts's law) when it's in the menubar to 
reduce pointing time.

http://www.nickgravgaard.com/windowlab/