Paul Jakma's patched version of zebra (still GPL of course).
zebra is a routing daemon implementing BGP, OSPF, RIPng, and OSPFv3,
among others, using a multi-process architecture.

This pkgsrc dir is intended to be used instead of net/zebra, and is
marked to conflict with it.  (The most recent released upstream source
for net/zebra is out of date and is missing some critical patches.)