Initial import of ap-ometer 1.2.0
mod_ometer is a web counter implemented as an apache C module. It uses gd 1.8.x
to generate its images, so no worries about the pesky Unisys/CompuServe
copyright whatever issue. Gd has some built-in fonts, but for real excitement
you can use any TrueType font you want for the counter.

In addition to allowing font choice, mod_ometer has all sorts of options for
customizing size, width and color of your counter as well as the ability to
output the counter as jpeg or png.

http://www.umich.edu/~umweb/software/mod_ometer.html

--
Testers are welcome :-)