Initial import of jabberd 1.4.2
jabber is an open XML protocol for the real-time exchange of messages and 
presence between any two points on the Internet. The first application
between any two points on the Internet. The first application of Jabber
technology is an asynchronous, extensible instant messaging platform, 
and an IM network that offers functionality similar to legacy IM systems
such as AIM, ICQ, MSN, and Yahoo.

--

This package needs some review by experienced people.
Please read the TODO file.