Initial import of maradns version 1.0.17
MaraDNS is a package that implements the Domain Name Service (DNS), an
essential internet service. MaraDNS is intended for environments where a DNS
server must be secure and where the server must use the absolute minimum
number of resources possible. MaraDNS was created in response to issues
people had with the DNS servers available in early 2001; and has the 
following design goals:

Security-aware programming. A DNS server needs to be secure. I have a number
of security features in the code, including:

   1. The code uses a special string library which is resistant to buffer
      overflows.

   2. The code, if started as root, mandates running as an unprivledged user
      in a chroot() jail.

Open-Source. The 1.0 release of the DNS server is public-domain code;
the 1.2 release will be released under a very liberal BSD-style license.

Simplicity. This DNS server has the minimum number of features needed to
correctly act as an authoritative and/or recursive name server.