Initial import of IlohaMail 0.7.11
IlohaMail (pronounced: e-lo-ha-mail) is a light weight yet full featured
multilingual webmail program that is easy to use and install. It runs on a 
stock build of PHP, and does not require databases (although database support 
is available) or the IMAP library (it is powered by a custom IMAP/POP3 
library).

http://www.ilohamail.org