Initial import of trickle 1.06
trickle is a portable lightweight userspace bandwidth shaper. it can run in
collaborative mode (together with trickled) or in stand alone mode.
trickle works by taking advantage of the unix loader preloading. essentially
it provides, to the application, a new version of the functionality that is
required to send and receive data through sockets. it then limits traffic
based on delaying the sending and receiving of data over a socket. trickle
runs entirely in userspace and does not require root privileges.

http://monkey.org/~marius/trickle/