Initial import of psh
The Perl Shell is a shell that combines the interactive nature of a Unix shell
with the power of Perl. The goal is to eventually have a full featured shell
that behaves as expected for normal shell activity. But, the Perl Shell will 
use Perl syntax and functionality for control-flow statements and other things.