Initial import of jabberd2 2.0.0-a4
Jabber is an open XML protocol for real-time exchange of messages and
presence notification between any two points on the Internet.  The first
application of Jabber technology is an asynchronous, extensible instant
messaging platform, and an IM network that offers functionality similar
to legacy IM systems such as AIM, ICQ, MSN, and Yahoo.

--
This package is the new release of the jabberd, has support for
openldap, cyrus-sasl2, ssl, ipv6 ...