Initial import of lmarbles, version 1.0.6. This closes pkg/20691.
I [the author, Michael Speck] always liked the game Atomix
and wanted to write a clone for Linux. But some guys had
this idea before me so I slightly changed the concept from
assembling molecules to creating figures out of single
marbles. Nevertheless, the basic game play is the same: If
a marble starts to move it will not stop until it hits a
wall or marble.

And to make it even more interesting there are obstacles
like one-way streets, crumbling walls and portals.

As Marbles is meant as a puzzle game you play against a
move limit and not a time limit. This way you have as much
time as you need to think.

Marbles has a menu to configure graphics (fullscreen,
animations), sounds and gameplay (player name, levelset,
difficulty).