Initial import of p5-DateTime-Timezone, version 0.19.
The DateTime::TimeZone modules provide a Perl interface to the Olson
time zone database.  Rather than using the database directly, we parse
the database files and turn them into a set of modules, one for each
time zone defined.  This allows for various optimizations in doing
time zone calculations.  This conversion is done with the script in
tools/parse_olson.

The Olson time zone database is the best available source for world
wide time zone information.  It is available from
ftp://elsie.nci.nih.gov/pub/.  A good starting point for information
on the database, and time zones in general, is
http://www.twinsun.com/tz/tz-link.htm.