Initial import of dfm by pancake
DFM is a file manager for Linux and other UNIX like Operating Systems. DFM 
is the abrvabation for Desktop File Manager. "Desktop" stands for the      
capability to place icons on the root window. The idea to write DFM came   
from OS/2. For a long time I had worked with OS/2. Using Linux I miss a    
desktop that provides easy launching programs, managing files and their    
association. I think only a program like the WPS can provide this. Some of 
the main features are:                                                     
     * Normal view, Detail view, Structure view                               
     * Icons may have any size                                                
     * placement of icons is free (Normal view)                               
     * every file have its own icon and startcommand                          
     * copy, move and link files                                              
     * The desktop is a special folder in the homedirectory - There the user  
       can place links to favourite programs or folders of the system         
     * every DFM-window have its own backgroundpixmap or color                
     * XV thumbnails are used as icons                                        
     * XDND drag support                                                      
     * Offix drag support