Initial import of zenity 1.2
This is zenity - a rewrite of gdialog, the GNOME port of dialog
which allows you to display dialog boxes from the commandline
and shell scripts. If you understand, things are just as they 
are. If you don't understand, things are just as they are.

--
Another package for the GNOME2 metapackage ready to import :)