Initial import of bitlbee 0.80
DESCR:

We are both console lovers. But it is annoying to have a few tty's open with 
chat things in them. IRC, ICQ, MSN, Jabber... For X there is Gaim, which 
supports many chatprotocols. Why wasn't there such a thing for the console?

The idea to port Gaim was easily thought of, of course. But we liked our IRC 
clients. And we used it the most, so we used it best. Importing it into the 
IRC client was a nice idea. But what if someone liked a different client. 
Then (s)he had to duplicate our work.

That's a shame, we thought. Doing work twice is pointless. So when Wilmer 
got the ingenious thought in his mind while farming, to create an IRC to 
other chatnetworks gateway, we were both so excited, that we started working 
on it almost immediately. And the result is BitlBee. 

--

http://www.lintux.cx/bitlbee.html