Initial import of cgdb 0.3.2
CGDB is a curses-based interface to the GNU Debugger (GDB). The goal of CGDB 
is to be lightweight and responsive; not encumbered with unnecessary features.

The interface is designed to deliver the familiar GDB text interface, with a 
split screen showing the source as it executes. The UI is modeled on the 
classic Unix text editor, vi. Those familiar with vi should feel right at home 
using CGDB.

--
http://cgdb.sourceforge.net/