Initial import of dpkg-1.9.21
This package contains the programs which handle the installation
and removal of packages on your system. The primary interface for
the dpkg suite is the `dselect' program; a more low-level and less
user-friendly interface is available in the form of the `dpkg'
command.

dpkg is for maintain the .deb packages of GNU Debian.

--
You can declare the DPKG_DB_PATH in your mk.conf to specify the dir
on which dpkg will install the database.

This is the first commit of the debian_linux packages for pkgsrc ...
will be continue soon ...