Initial import of festival-gaim, version 0.64:
festival-gaim is a plugin for the Gaim IM client that synthesizes incoming
messages using voice.  The synthetizer used is festival.

I'm importing it in pkgsrc-wip since the plugin makes gaim crash just after
enabling it.  I've got in contact with the author, but until he replies, I'll
leave it here.  It could be interesting to try this on Linux (the original
development platform) and see if it works (assuming all dependancies compile
properly ;).