Initial import of ish (icmp shell)
--
I'm tired today...this is the last one ;)...but I will back again
--


ICMP SHell is a telnet-like protocol. It allows users to connect to a
remote host and to open a shell using only ICMP to send and receive
data. ICMP Shell war written in C for the UNIX environment.