Entropy, a Freenet-like network node written in C (rather than Java 1.4). Still very much beta quality, but very fast and efficient compared to Freenet.  Also supports Freenet's client protocol (FCP).
Somewhat appropriate for today:  Declare independence for your data.  8-)

===== From DESCR:

ENTROPY stands for Emerging Network To Reduce Orwellian Potency Yield and as
such describes the main goal of the project.

ENTROPY is developed as a response to increasing censorship and surveillance
in the internet. The program connects your computer to a network of machines
which all run this software. The ENTROPY network is running parallel to the
WWW and also other internet services like FTP, email, ICQ. etc.