Initial import of drwright 0.17 - A computer break reminder.
DrWright is a program that reminds you to take regular breaks from your
computer. Its main focus is being usable and to simply work.

DrWright monitors keyboard and mouse activity and displays a small meter
in the notification area on your panel. When it starts blinking, you know it's
time to take a break. If you don't, it will lock your screen to force a break.