Initial import of Cyrus Imapd 2.2.1-BETA.
The new features include Virtual Domain support, NNTP
server support, IPv6 support, Compiled Sieve scripts, an improved
configuration subsystem, and many more visible and invisible changes.

IMAP (Internet Message Access Protocol) is an Internet standards-track
protocol for accessing messages (mail, bboards, news, etc). The Cyrus
IMAP server differs from other IMAP server implementations in that it
is generally intended to be run on "sealed" servers, where normal users
are not permitted to log in. The mailbox database is stored in parts of
the filesystem that are private to the Cyrus IMAP system. All user
access to mail is through the IMAP, POP3, or KPOP protocols.