CPAN module Class::MethodMaker:
This module solves the problem of having to write a bazillion get/set
methods that are all the same.  The argument to 'use' is a hash whose
keys are the names of types of generic methods generated by MethodMaker
and whose values tell method maker what methods to make.  More
precisely, the keys are the names of MethodMaker methods (methods that
write methods) and the values are the arguments to those methods.